docs: replace Harness example and fix links in README (#1820)
Fixes #1817. This is a quick PR to replace the Harness example in the
Ops README by the latest (Scenario) unit test that comes from
`charmcraft init`. We're planning to rework the README at a later date
(see https://github.com/canonical/operator/issues/1655) but I think this
is a reasonable temporary solution.

I've also fixed a couple of links to the Ops reference docs.