docs: add httpbin charm from Charmcraft as an example charm (#1743)
In Charmcraft 4.0, we're planning to remove the `simple` profile of
`charmcraft init` (see
https://github.com/canonical/charmcraft/pull/2295). This PR adds
`simple`'s charm as a demo of Ops, so that the charm isn't lost.

I took the charm that was generated by `charmcraft init` and
changed/removed placeholder comments & files that were more suitable for
a template charm. I also made sure that the charm is formatted correctly
for the Ops repo and passes our linting.

Within the charm directory, I tested that the following work:

- `tox -e static` - For now, this runs static checks on _src/charm.py_
only. When we make future improvements including switching to uv (see
#[1805](https://github.com/canonical/operator/issues/1805)) we could
expand `static` to cover the testing code too.

- `tox -e unit` and `tox -e integration`

- Manual pack and deploy, as documented in _examples/README.md_ and (for
the time being) _README.md_.

I also added the charm to our CI job that runs unit tests for the
example charms.

---

Follow-up items in future PRs:

- https://github.com/canonical/operator/issues/1655
- https://github.com/canonical/operator/issues/1805