fix: do not return current unit in a mocked peer relation (#1828)
This PR filters out the current unit from being returned by Scenario in
the case of a peer relation, to align with Juju's behaviour. A follow up PR
will fix the underlying buggy behaviour that leads to Scenario's peer
relations dict being populated with an entry for the current unit.