fix: testing.PeerRelation properly defaults to no peers (#1832)
The intention of Scenario's `PeerRelation` is that it defaults to no
peers (the charm only has the one unit). In addition,
`PeerRelation.peers_data` should only contain unit data for the *peers*,
not for the unit that is under test - that is found in
`local_unit_data`.

We were breaking these rules when creating a `PeerRelation` without an
explicit `peers_data`. We would add in the default Juju relation data -
but that was already present in `local_unit_data`, and we used `0` as
the unit ID. In most cases, `0` is the unit being tested, so this data
was mostly ignored and unable to be reached, although it was the
underlying cause of the bug fixed in #1828. The worse situation is if
someone provides a unit ID to `Context`, other than 0, and has a
`PeerRelation` with default `peers_data`, the charm would 'magically'
have two units.

The main fix here is that `peers_data` should be the empty dictionary by
default. If charmers want peers, then they need to pass in an
appropriate dictionary (and handle the default Juju keys themselves if
relevant).

Since it's very simple to get `peers_data` wrong and put in data that
should be in `local_unit_data`, the PR also adds a consistency check
that verifies that everything is ok and suggests a fix if not.

---------

Co-authored-by: Dave Wilding <tech@dpw.me>