fix: in meter-status-changed JUJU_VERSION is not set (#1840)
Despite the
[documentation](https://documentation.ubuntu.com/juju/3.6/reference/hook/)
saying that `JUJU_VERSION` is set in the environment for all hooks, it
is not for the `meter-status-changed` or `collect-metrics` events.

Restore the older behaviour (prior to #1313) where if `JUJU_VERSION` is
not available we set a version of `0.0.0`. We can drop this once ops
requires Juju 4 and above.

In addition, in the restricted context the `is-leader` hook tool is not
available, which means that collect-status fails (before emitting any
events, because it can't figure out which events to emit). We could just
do the unit status in this case, but it seems simpler to just skip
collect status in this case (which is only for two events no-one should
be using any more).

Finally, the PR adds `meter-status-changed` to the list of restricted
context events, which means that we won't run any deferred events during
those hooks (as with `collect-metrics`), and will error out cleanly if
`use_juju_for_storage` is used. Not emitting deferred events is a
breaking change, but definitely should be done - it's extremely likely
that the handler would crash if unexpectedly running in a restricted
context. Similarly, for the `use_juju_for_storage` case, the charm would
have less gracefully crashed anyway, when ops tried to use storage
itself.

Fixes #1836