feat: provide a method to create a testing.State from a testing.Context (#1797)
A new `from_context` class method is added to `testing.State`, which
returns a `State` object that is populated by the metadata, config, and
charm class.

* The config is set to the default values (as defined in
`charmcraft.yaml`)
* A `testing.Relation`, `testing.PeerRelation`, or
`testing.SubordinateRelation` is added for each relation in
`charmcraft.yaml`'s `peers`, `requires`, and `provides` fields. The
relation has the endpoint and interface set, but no data. Note that
relations already automatically have a `Network` defined. Note that this
simulates `integrate` having already run.
* A `testing.Container` is added for each container in
`charmcraft.yaml`'s `containers` field. The container's `can_connect` is
set to `True`, unlike the default `False` with `Container(name)`; this
simulates the container having already been created
* A `testing.Storage` is added for each storage in `charmcraft.yaml`'s
`storage` field
* A `testing.StoredState` is added for each `ops.StoredState` object
that is a class attribute of the charm

In some cases, this will be a more useful starting point for a `State`
than a blank `State` as provided by `State()`.

The biggest issue is that because the `State` is immutable, it is then
awkward to do `dataclasses.replace` calls to insert data and do other
modifications. To help with this, extra/non-default components can be
passed in, just as they would be to the regular `State()`. For example:

```python
relation = testing.Relation('db', local_app_data={'foo': 'bar'})
state_in = testing.State.from_context(ctx, leader=True, relations={relation})
```

Fixes #1454

---------

Co-authored-by: James Garner <james.garner@canonical.com>