fix: only provide the units belonging to the app in Relation.units (#1837)
The previous PR for #1109 wrongly added the `REMOTE_UNIT` unit to all
relations, but it should only be ones where the app matches (and in
particular, is never for peer relations). This fixes that, and adds an
extra unit test for verification.

To be extra sure, an integration test is added that verifies that the
correct units are returned. The test/integration conftest was quite
specific to tracing, so this also makes it more generalised.

I'd like to have a few integration tests for ops, and the door has been
opened with tracing, but I'm happy to discuss this if people feel it's
not worth doing for this case (I had to more-or-less make it anyway to
verify the fix). It does form quite a significant portion of the changes
here.

Fixes #1109

---------

Co-authored-by: James Garner <james.garner@canonical.com>