docs: be consistent with recommending self.app and self.unit (#1856)
This PR starts work on #1783. In the interests of making progress, I've
kept the scope focused on docs related to `self.app` and `self.unit`.

I've updated several docstrings to hopefully steer people away from
using `Model.app` and `Model.unit` when `CharmBase.app` and
`CharmBase.unit` are appropriate. While I was doing this, I also updated
some of the wording to be more consistent.

Main updated parts:
- Descriptions of
[`Application`](https://ops--1856.org.readthedocs.build/en/1856/reference/ops.html#ops.Application),
[`Model.app`](https://ops--1856.org.readthedocs.build/en/1856/reference/ops.html#ops.Model.app),
and
[`Model.get_app()`](https://ops--1856.org.readthedocs.build/en/1856/reference/ops.html#ops.Model.get_app)
- Descriptions of
[`Unit`](https://ops--1856.org.readthedocs.build/en/1856/reference/ops.html#ops.Unit),
[`Model.unit`](https://ops--1856.org.readthedocs.build/en/1856/reference/ops.html#ops.Model.unit),
and
[`Model.get_unit()`](https://ops--1856.org.readthedocs.build/en/1856/reference/ops.html#ops.Model.get_unit)

I didn't find any places that needed updating in the markdown source.