docs: update HACKING.md with how to bump Ops version in Charmcraft profiles (#1872)
In https://github.com/canonical/charmcraft/pull/2312 we're switching
Charmcraft profiles to use `pyproject.toml` for dependencies -
specifically Ops - and include `uv.lock` in generated charms. After
discussion within Charm Tech, we decided that we should update the deps
and lockfile whenever there's a major new feature of Ops or we resolve a
dependency issue.

This PR updates the Ops release process in HACKING.md with instructions
for doing that.