fix: if an event ends with _Abort(0) tests should behave as if it ended successfully (#1887)
If ops code raises `_Abort(0)`, then the event completes successfully as
far as Juju is concerned. The fact that ops has shortcut things to jump
right to exit isn't really relevant from the Juju point of view.

This means that for Scenario tests, we should treat `_Abort(0)` as a
normal successful event completion, because it's simulating the events
from the point of view of Juju.

Raising `_Abort` with any other exit code still ends up as
`UncaughtCharmError`, as that indicates that something has gone wrong,
and from the Juju perspective changes will be rolled back and the charm
will end up in error status.