docs: Use config and action classes in the Kubernetes tutorial (#1891)
1. In the config chapter, we add a config class (with the single option)
and add validation for it. I've put the loading into the pebble layer
property, as that's (initially) the only place that it's used and it
seemed cleanest for the tutorial. The invalid value handling is then
moved to a small except block in the update and restart code.
2. In following chapters, continue with the approach for config, but
when we move to collect-status also add validation of the value there.
In the observability chapter, also handle getting the config for
metrics.
3. In the actions chapter, we add an action class (with the single
parameter of the action), and use it when loading the parameters. This
is a very minimal change, but demonstrates the use of the feature, and
encourages continuing with it when there are more complex actions. This
uses `errors='fail'` because that is appropriate for actions with most
charms and is cleaner code -- in this specific case, Juju will prevent
any invalid value, so there will never be a failure.

pre-commit also fixed some unrelated trailing whitespace - you probably
want to review with whitespace changes removed.


[Preview](https://ops--1891.org.readthedocs.build/en/1891/tutorial/from-zero-to-hero-write-your-first-kubernetes-charm/index.html)

---------

Co-authored-by: Dave Wilding <tech@dpw.me>
Co-authored-by: James Garner <james.garner@canonical.com>