fix: if self.app is not actually set avoid a new crash location (#1897)
`Model.app` should always be set (and code comments explicitly state
that). However, [a Juju bug where JUJU_REMOTE_APP is not
set](https://bugs.launchpad.net/juju/+bug/1960934) means that is not
always true.

We work around this
[elsewhere](https://github.com/canonical/operator/blob/a15646475eab8d243ef03c31d0e76ccb02941050/ops/charm.py#L570),
so we should do it when potentially adding in the departing unit as
well. Once the upstream issue is solved, we should be able to remove
this.