chore: minor type checking fixes (#1903)
Fix a variety of small type annotation issues.

These are found with `ty` and missed by `pyright` but should be fixed
both because they are issues and because it would be good to be more
prepared for a stronger type checker in the future. (Note that we can't
move to `ty` yet because (a) it's marked as not ready, and (b) there are
a bunch of false positives still.)

---------

Co-authored-by: James Garner <james.garner@canonical.com>