docs: replace broken link in testing explanation (#1913)
The files in the promtheus-k8s repository have moved around, breaking
the example link the testing explanation doc.

This PR replaces that link, but rather than update it to the new
prometheus-k8s location, uses the ubuntu-manpages charm instead. That
charm should be more stable, the tests are much simpler (which better
matches this specific part of the docs, which are explaining the
basics), and the charm *only* uses Scenario, unlike prometheus-k8s,
which has a mix of Harness and Scenario.

Fixes #1910