ci: add SBOM generation and secscan workflow (#1906)
This repo now has too many artifacts (3 PyPI packages each in two
different formats each with two maintained versions) to manually request
the SBOM generation and secscan. This PR adds configuration files for
the [sbomber](https://github.com/canonical/sbomber) tool to automate
those processes.

The tool can be run manually against the configuration files, but there
is also a workflow that will do this automatically. It is currently set
to be manually triggered, but the intention is that once it is verified
to work it will also be added to the publish workflow, to completely
automate the process.

The secscan tool can only be run in the Canonical network, so the
workflow is set to use a hosted runner. This prevents testing the
workflow in my fork.