fix: only add the remote unit for departed and broken relation events, fix ordering (#1918)
We inject the `JUJU_REMOTE_UNIT` unit into the relation's unit so that
the databag is available during relation-departed events. When this was
added in #1364 the thinking was that it was cleaner to just always add
that unit rather than check for the specific events. However, that was
not correct: this means that the unit may be added before the data is
available (in relation joined, for example).

This PR changes the injection so that it only happens in relation
departed and relation broken events.

In addition, we inserted the unit into the relation *after* setting up
the relation's `.data`, which meant that the data would not be
accessible for the inserted unit. The PR adjusts the ordering so that
the inserting is done first.

Fixes #1917