feat: release automation script (#1855)
This PR automates the release process according to our discussion in
Frankfurt, see
[here](https://docs.google.com/document/d/1fkoqp1A2SMiOnwHubOJVRkuD1GpvcYxSDmfExYIZO88/edit?tab=t.0#heading=h.ukrjinktm94z)
(internal link) for more details.

Dependencies:

- ENV var `GITHUB_TOKEN` set
- The `git` command is available locally
- `pygithub` (managed by tox; if you want to run the script directly for
testing purposes, you need to have it).

The script does the following:

- Runs locally with minimal dependencies (`pygithub`, installed by tox,
and `git` command)
- By default, release from the main branch, but other branch is
supported.
- Release: `tox -e release`. This detects version from the latest tag
and suggests a new version for user confirmation. It can be overriden
with user-provided value.
- The script can detect open "update charm pins" PR and asks the user to
merge those first.
- The script asks for a human-written title and a introductory paragraph
for the release.
- Release note and `CHANGES.md` are formatted (for example, drop the
GitHub usernames)
with categorizations accordingly
- The script automates version number changes in various files and
creates a PR.

@tonyandrewmeyer @james-garner-canonical please have a look because I
think according to the order it's going to be you guys doing releases
soon.

Done:

- [x] Test the script by doing a real release.
- [x] Update docs dependencies as separate process/PR - no need to do
anymore since we moved everything to uv with lock.

---------

Co-authored-by: James Garner <james.garner@canonical.com>