fix: add the remote unit to Relation.data but not Relation.units (#1925)
In #1364 (to address #1109) we added the departed unit to
`Relation.units` so that charms could access the relation data in the
relation-departed event.

This was followed up by #1897, to fix an issue where the app was not
available, and then #1918, to add the unit before populating
`Relation.data`, and to only add the unit for `relation-departed` and
`relation-broken`, rather than all events.

A problem that the original solution introduces is a subtle behaviour
change in what `Relation.units` means. It's not explicit in the
documentation, but previously it was always the same as the
`relation-list` output. Changing which units are included could cause
issues for holistic or reconcile charms that would now see the departed
or broken unit as still part of the relation, until the event after
relation-broken.

In this PR, the implementation is adjusted so that rather than adding
the remote unit to `Relation.units`, it is added to `Relation.data`.
This restores the original meaning of `Relation.units`, but still allows
charms to access the data of the departed or broken unit - but the charm
will need to do so explicitly (via `event.relation.unit` for example)
rather than looping through `Relation.units`.

The fix from #1897 and the 'only for departed and broken' parts of #1918
are kept (the other part of #1918 is no longer relevant).

The tests that were added to validate the "adding an extra unit to
`Relation.units`" behaviour are removed, but the test for validating
that the data is accessible is kept (and adjusted to explicitly request
the data).