ci: fixes for the sbom and secscan workflow, and trigger it on publishing (#1916)
Connect the publishing workflow to also trigger SBOM generation and
secscan for the new artifacts. Also adds instructions to the release
process to copy the resulting reports to the correct location.

Also fixes a few issues with the existing workflow:
* Explicitly install the libapt-pkg-dev package, which is required to
install the Python apt package.
* Upload the reports into separate wheel and sdist locations, to avoid
clashes.
 * Remove the 2.23/7.23 versions -- those should be in the 2.23 branch.
 * Remove an old configuration setting from the wheel config.

Also makes use of the new ability of sbomber to automatically detect the
version number, so that the manifest files don't need to be updated for
each release.

For ops 2.23, ops-scenario 7.23, and ops-tracing 2.23, new releases will
continue to use manual processes.