feat: add security event logging (#1905)
Adds logging of security events, as specified in
[SEC0045](https://docs.google.com/document/d/1nInWP9pEEhloKMfgzDsd4Pub4530OxQcKKh716Gvxfk/edit?usp=drivesdk).

A new `_security_event` helper is added to `ops.log`. Security events
must be logged as structured data, so the helper does this, and provides
some fields common to all events (a timestamp, the model UUID, and a
type of "security") as well as the provided event, level, and
description.

This method is private and not available to charms. There was almost no
interest in having a public method, so that seems the best choice for
now. If we decide to offer this functionality to charms in the future,
that only requires removing the leading `_`, adding the method to the
`ops` namespace, and adding back the other event types (basically
reverting 139f4de3317f82e7563a42e78a42ecec77845d8b).

Logged security events are:
* `sys_crash`, if the charm exits because of an uncaught exception.
* `sys_restart`, if the unit's `reboot` command is executed.
* `authz_fail`, if any hook command fails with error text containing
"access denied" and some other similar errors (OP067 has a full list of
the error cases).

See also [OP067 - Security Event Logging for Charm
SDK](https://docs.google.com/document/d/1VsxGDxuUOo6A3PfT_qzFssR3qTOJInrHPORitnbtiDk/edit?usp=sharing)

As a drive-by (previously related, but not any more): the `str()` of a
Pebble basic identity will always mask out the password.