docs: for 2.x, remove .html extensions (#1954)
In https://github.com/canonical/operator/pull/1951 we started updating
the doc starter pack for the `latest` version of the Ops docs. That
included switching to "clean" URLs (without .html).

This PR switches the `2.x` version of the docs to clean URLs too. I
don't think it's worth doing further starter-pack related changes in
2.x, as we don't plan to update the 2.x docs frequently. But switching
to clean URLs is helpful to avoid 404s when switching between doc
versions.

Example of how to produce a 404:

1. Visit
https://documentation.ubuntu.com/ops/latest/howto/manage-charms/
2. Use the version switcher to switch to the 2.x docs

The resulting page
https://documentation.ubuntu.com/ops/2.x/howto/manage-charms/ is not
found. I _might_ be able to work around this in the RTD redirect
configuration, but it would be hacky. It's better to switch the 2.x docs
to the `dirhtml` builder.