docs: for 2.x, fix unstyled error pages (#1973)
Error pages in the 2.x docs are broken. For example:
https://documentation.ubuntu.com/ops/2.x/foo

I previously tried several things to fix it. Last week I found the cause
and fixed it for the latest version of the docs
(https://github.com/canonical/operator/pull/1969). This PR ports the fix
to the 2.x version of the docs.