chore: add refactor, perf, and revert to the changelog/release note categories (#1974)
The release notes and changelog currently ignore any commits that are
`refactor`, `perf`, `revert`, or `chore`. We are deliberately ignoring
'chores', as those are deemed uninteresting and not relevant to
consumers of the packages.

There are two cases for "revert". One is reverting a commit that is also
in the same release; in this case we should remove both the original
commit and the reverting commit to have a cleaner list of changes -
doing this automatically is tricky, so having the revert listed in the
changes makes it simpler to do this in the human review. The other case
is reverting a change from a previous release: consumers of the packages
should definitely be told about these.

Performance improvements are typically interesting to users, so it makes
sense to mention those in the release notes and changelog. In extreme
cases, the changes might make it possible to use a feature that wasn't
feasible previously. Even in minor cases, it helps to show that we care
about performance.

Refactoring is arguably not interesting to users, if there are no
performance changes and no changes to the public interface of the
packages. However, we decided as a team that we should include these as
well, at least for now (and we have a very small number of `refactor`
commits, so it doesn't make much difference in practice).