chore: remove the version check for importing ops-scenario in ops.testing (#1980)
Since we pin the version of ops-scenario to an exact release, we no
longer need to check if the installed version of Scenario is the right
one. If someone has separately installed `ops-scenario` < 7 (and not
`ops[testing]`, which would replace it) then the import will fail on
names that don't exist (like `layer_from_rockcraft`).

This simplifies the ops.testing namespace setup, and removes the
first-level dependency on importlib-metadata (opentelemetry-api still
requires it, so it moves to a second-level dependency).

Also remove compatibility code in the Scenario imports for older
versions of Scenario, which is no longer required.