ci: adjust permissions block in publish workflow (#1984)
This fixes the two open code scanning alerts. It's unclear how the
failing workflow managed to get merged despite the failures (it doesn't
appear to be due to a difference in Zizmor version, or an override, and
the history in the scanning alerts does not seem correct).

This moves the permissions (particularly attestations:write and
id-token:write) to the build-and-publish job, rather than applying to
the entire workflow. In practice, this means that the secscan job
doesn't have those permissions.