ci: enable doctests (#1991)
In #1975, we'd like to include a couple of docstrings. This PR clears
the way for that, as well as any other doctests we wish to use in the
future.

The PR tweaks some docstrings that are in private methods, so not in in
our reference documentation. In addition:

* `ActionEvent.set_results` changes from a block of `>>>` text to a set
of bullet points. If this was doctest'ed then it'd actually call the
method, but doing that with `set_results` itself would be complicated.
This seems just as clear to me.
* `pebble.Client.exec` has a couple of lines to disable doctest. This
could be a doctest, but that would require a real Pebble, which we
currently split off into a different set of tests. This is a bit ugly,
but putting `doctest: +SKIP` on most of the lines is uglier (in my
opinion), and I don't think there's a cleaner way to skip a single
docstring.

---------

Co-authored-by: Dave Wilding <tech@dpw.me>