docs: change some best practices to tips (#2001)
We're using "best practice" blocks to mean "required for public listing
on charmhub". That implies that they should be items that are reasonable
to review & resolve at that time, rather than earlier in the charm
development process. We're generally using "tip" blocks for suggestions
that occur earlier.

This PR changes some "best practice" blocks to tips:

* `logger.info(f"foo {bar}")` rather than `logger.info("foo %s", bar)`
(in practice this also doesn't tend to make a lot of difference in the
charming case, although it is more correct)
* Absolute paths in `subprocess` calls (this is also less relevant in
the charming context than in general Python code)
* Avoid running subprocesses via the shell.

[More
context](https://github.com/tonyandrewmeyer/charmhub-listing-review/issues/3).

---------

Co-authored-by: Dave Wilding <tech@dpw.me>