docs: improve intro to ops.testing reference (#2023)
Resolves #1826. This PR makes a few small improvements to the intro of
the `ops.testing` reference:

- Mention that how-to guides have info about testing each feature
- Link to the unit testing how-to guide and the testing explanation
(moved from the end of the intro)
- Mention the term "unit tests" much earlier
- Move the "previously known as 'Scenario'" part to the end of the
intro, as Scenario is already mentioned in the title

**[Preview
doc](https://canonical-ubuntu-documentation-library--2023.com.readthedocs.build/ops/2023/reference/ops-testing/)**

---

Not implemented:

- In [#1826](https://github.com/canonical/operator/issues/1826), I
mentioned possibly adding a link from
[`Relation`](https://documentation.ubuntu.com/ops/latest/reference/ops-testing/#ops.testing.Relation)
to the how-to guide. But why only from `Relation`? I'm adding the new
info right at the beginning of the intro, so let's first see how well
that works.

- I also mentioned using full URLs for links, so that the links work in
docstrings. We don't actually need to do that here. The docstring of
`ops.testing` comes from `ops/testing.py` not the intro of
`docs/reference/ops-testing.rst` - that intro is purely for the docs.