add first draft of new machine charm tutorial (#1999)
This PR starts work on a new machine charm tutorial. The purpose of the
new tutorial is to teach modern charming practices with a real (but not
complex) workload.

**The new tutorial is not ready to be used by charmers yet.** But early
feedback is extremely welcome. To avoid dropping a huge PR for the team
to review, this is the first PR in a sequence. Subsequent PRs will fill
in the TODOs and make any changes coming from this initial review. Also,
the tutorial won't be ready to use until Charmcraft 4 and
`charmlibs-apt` are released.

Focus for reviewing this PR:
- Overall structure and topics covered in the tutorial
- Code review, including unit tests

**[Preview
doc](https://canonical-ubuntu-documentation-library--1999.com.readthedocs.build/ops/1999/tutorial/write-a-machine-charm-for-tinyproxy/)**
(ultimately the tutorial will replace "Write your first machine charm" -
it won't be a separate doc)