feat: expose the Juju hook context in ops.JujuContext (#1996)
Renames `ops.jujucontext._JujuContext` to `JujuContext` and exposes it
in the top-level `ops` namespace.

The `from_dict` method is kept private (by renaming it to `_from_dict`).
This is still used internally, and is the most forgiving method for
instantiating a `JujuContext`. A new `from_environ` method is added,
which strictly validates that the correct variables are in the
environment (based on the [Juju
documentation](https://documentation.ubuntu.com/juju/3.6/reference/hook/)
and manual validation).

(The majority of the changes in terms of lines are from those two
renames).

`JujuContext` now requires keyword arguments (presumably we would have
done this originally, but were restricted to Python 3.8 at the time).

`JujuContext` has a few more attributes than `_JujuContext`:
* `availability_zone`
* `hook_name`
* `machine_id`
* `principal_unit`
* `storage_index` (this was overlooked in the spec)

The `JujuContext` reference docs
([Preview](https://canonical-ubuntu-documentation-library--1996.com.readthedocs.build/ops/1996/reference/ops/#ops.JujuContext))
are tweaked slightly to be aimed at charmers (or people writing charming
APIs/frameworks) rather than developers working on Ops itself. There are
no additional how-to guides or explanation docs, as we do not intend for
this class to typically be used by charmers (everything charmers need
should be available via the framework). The follow-up PR that adds
`ops.hookcmds` might add a general 'writing your own charming API' doc
(very much to be decided).

The spec has an `event_name` property being added to `JujuContext`,
however, when implementing this I realised that it's essentially just
`.hook_name`, so there doesn't seem any point in adding that.

Scenario is switched over to use `from_environ`, as we are in total
control over the environment in that case, and this exercises the new
code significantly.

There are two small adjustments to Scenario that arose from this work:

* A new consistency check for relation events, verifying that they
provide the remote unit. There are existing warnings at (test) runtime
if the remote unit is implicitly selected -- these now also produce
consistency check warnings. At (test) runtime if there is no remote unit
to be selected, the test would previously have continued after warning,
but that scenario is not possible in Juju, so now raises a consistency
error.
* The mocked `os.environ` that Scenario manages now correct uses dashes
for `JUJU_DISPATCH_PATH` and `JUJU_HOOK_NAME`.

See
[OP069](https://docs.google.com/document/d/1HJSFVxDSDVYatVcvv4k13nBMvohb9sdgWbRt2pGOiyk/edit?usp=sharing)
for more details.

---------

Co-authored-by: Dave Wilding <tech@dpw.me>