fix: `_MockModelBackend.relation_get` will return a copy of the relation data (#2052)
Previously, `_MockModelBackend.relation_get` returned the same `dict`
object associated with the `Relation` objects it managed, rather than a
copy. This meant that when `ops.RelationDataContent` was instantiated,
it was passed this `dict` to use as its local cached copy. This meant
that our proxy for Juju's relation data store and the local cache were
the same exact `dict` object in testing.

At runtime, when relation data is updated, first `relation_set` is
called to update the Juju data, and then the local cached copy is
updated to match. During testing, this hid a subtle bug in our mock
`relation_set` logic -- entries were not deleted as expected, but this
error was 'corrected' by `RelationDataContent`'s update of its local
cached copy (the same `dict`) deleting the entries.

This PR updates `MockModelBackend.relation_get` to return a copy of the
`dict` instead. This means that `_MockModelBackend.relation_set`'s logic
is now exercised as expected (and has been corrected to handle removals
as a result -- tests fail when we return a copy without making this
change to the tests).

Resolves #1638.

---------

Co-authored-by: Ben Hoyt <benhoyt@gmail.com>