docs: update referenced examples for managing interfaces (#2068)
In "How to manage interfaces" we have some links to a repo that
demonstrated the code shown in the doc. But over time some of the links
have broken. In a team discussion we decided to remove the references
and keep the links to examples that are in active use.

This PR updates the references. In the future, we may need to make more
substantial changes as interfaces move to the
[charmlibs](https://github.com/canonical/charmlibs) repo.

**[Preview
doc](https://canonical-ubuntu-documentation-library--2068.com.readthedocs.build/ops/2068/howto/manage-interfaces/)**

Drive-by: pre-commit cleaned up some whitespace. I suggest reviewing
with whitespace hidden.