chore: improve release process instructions, mainly related to secscan (#2067)
While doing the release today, I manually ran the "SBOM and secscan"
workflow because I didn't realise that it would run as part of the
Publish workflow. This PR updates the instructions in HACKING.md to
explain where to find the artifacts (without manually running the
workflow).

Drive-by: Pre-commit tidied some of the whitespace and I did a little
more tidying. It might be easiest to hide whitespace when reviewing.