docs: switch to makefile for building the docs (#2073)
This PR switches the doc build process to the makefile from the [starter
pack](https://github.com/canonical/sphinx-docs-starter-pack/tree/main/docs).
In the future, we'll be able to use the makefile for more things than
building the docs, e.g., running the new spellchecker. But I haven't
added the required config files yet. I'll do that in a follow-up PR.

I had originally hoped to use the makefile as-is from the starter pack.
However, I couldn't reconcile two things:

1. We need `ops[testing,tracing]` installed from source during the doc
build.
2. I want `make -C docs html` to work without any prior commands (for
consistency across projects).

The solution I landed on is to modify (only) the part of the makefile
that creates the virtual environment. I'm using uv to create a venv,
then install the docs requirements and `ops[testing,tracing]` into the
venv.

In case these changes to the makefile get overwritten in the future,
I've added a check in `docs/conf.py` to make sure that Ops is installed.

---

To summarise:

| Old way of building docs | New way of building docs |
|--------------------------|--------------------------|
| `tox -e docs`            | `make -C docs html`      |
| `tox -e docs-live`       | `make -C docs run`       |