feat: add a low-level API for the Juju hook commands (#2019)
Adds a new module, `hookcmds`, that has methods for calling the Juju
hook commands. The intention is that there is one method for each hook
command (other than ones that are explicitly deprecated), and that all
the functionality that makes sense is exposed, even if it's not used in
the Ops framework. The arguments and return types are Pythonic.
Basically: the Jubilant approach to the hook commands.

This PR adds ops.hookcmds (and tests and reference documentation) but
does not change `ops.model.py` to use it. I'll do a follow-up PR right
after this one is merged that adjusts `_ModelBackend` to be a light
wrapper around `ops.hookcmds` - it seems like this will make reviewing
simpler.

I've also separated out adding integration tests for `hookcmds` - I do
believe that this is a valuable addition, but it will be simpler to
review separately, and it's not essential.

See the
[spec](https://docs.google.com/document/d/1HJSFVxDSDVYatVcvv4k13nBMvohb9sdgWbRt2pGOiyk/edit?usp=sharing)
for more details.

---------

Co-authored-by: Ben Hoyt <benhoyt@gmail.com>