fix!: ensure that the testing context manager is exited when an exception occurs (#2117)
Currently, if `testing.Context` is used as a context manager, and an
exception occurs during the event, proper cleanup is not done. This is
noticeable in two ways in particular:
* Exceptions aren't wrapped in `UncaughtCharmError`.
* The environment is not cleaned up. This can manifest in a cascading
failure of tests if `OPERATOR_DISPATCH` is left set.

We currently manually call the `__exit__` function of one of the context
managers when wrapping up. However, if an exception occurs, we never
reach that point, and the context manager is left un-exited (you'll also
often see errors when exiting Python, because it doesn't like being left
in that state). The PR fixes this by also calling `__exit__` if exiting
the `Context` context manager, and it's handling an exception.

Many thanks to @james-garner-canonical for helping figure out where
things were wrong wrong and a good way to fix this.

Fixes #1926