refactor: use ops.hookcmds in _ModelBackend (#2116)
This PR removes the subprocess calls from model.py's `_ModelBackend`,
replacing them with using the new ops.hookcmds package.

Other changes:
* Fix the `_NetworkDict` definition (several fields were missing). This
introduces use of `NotRequired` so it also moves into the
`TYPE_CHECKING` block.
* `network-get` has the `--format=json` argument first, as the other
commands do, and as the documentation suggests.
* `opened-ports` no longer raises if there's an unexpected protocol,
because model only warns about this, so we need the lower-level to
accept other protocols for backwards compatibility.
* `relation-list` returns a `str` when `--app` is used, and a `list[str]
`otherwise.
* Adjust the order of `--app` in `relation-set` to match `relation-get`
(this is just for more consistent looking commands).
* Fix a comment typo that @benhoyt noticed in the previous PR.
* Adds a missing default argument in `secret-revoke`.
* A lot of the network fields have tests for them being missing, so I'm
assuming that might happen (I think it's actually in very old Juju, but
it seems good to be conservative) and loading the data class is adjusted
to handle that.
* Drop support for testing on Windows (this was simpler than adjusting
the batch scripts, and we already gave up on Windows).
* Drops support for using `relation-get` on Juju 2.7 and 2.8 where
there's no `--app` argument (and the corresponding tests). I believe our
minimum supported Juju is 2.9 with Ops 3.x.

---------

Co-authored-by: James Garner <james.garner@canonical.com>
Co-authored-by: Dima Tisnek <dima.tisnek@canonical.com>