ci: add automated doc checks (and related starter pack updates) (#2099)
This PR does a few things related to the [documentation starter
pack](https://github.com/canonical/sphinx-docs-starter-pack):

- Makes sure that the spell checker can be run with `make -C docs
spelling`, makes sure that the spell checker _passes_, and adds a
workflow to run the spell checker on PRs & pushes.

- Configures the link checker with URLs that are known to fail, and adds
a manual workflow to run the link checker. Manual for now because the
link checker takes a long time and produces unexpected errors (see my
comments for more detail).

- Adds the update script from the starter pack.